Force size_t.

+++ tests/cpp/common/test_hist_util.cu
@@ -52,7 +52,7 @@ TEST(HistUtil, SketchBatchNumElements) {
   size_t constexpr kCols = 10000;
   int device;
   dh::safe_cuda(cudaGetDevice(&device));
-  auto avail = dh::AvailableMemory(device) * 0.8;
+  auto avail = static_cast<size_t>(dh::AvailableMemory(device) * 0.8);
   auto per_elem = detail::BytesPerElement(false);
   auto avail_elem = avail / per_elem;
   size_t rows = avail_elem / kCols * 10;
